+++ app.js
@@ -1,9 +1,8 @@
-var createError = require('http-errors');
-var express = require('express');
-var path = require('path');
-var cookieParser = require('cookie-parser');
-var logger = require('morgan');
-const pg = require('pg');
+const createError = require('http-errors');
+const express = require('express');
+const path = require('path');
+const cookieParser = require('cookie-parser');
+const logger = require('morgan');
 const cors = require('cors')
 const app = express();
 
@@ -14,8 +13,9 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
-var indexRouter = require('./routes/index');
-var usersRouter = require('./routes/users');
+const indexRouter = require('./routes/index');
+const usersRouter = require('./routes/users');
+const roomRouter = require('./routes/room')
 
 const socket = require('socket.io')
 const server = app.listen(3001, () => console.log('Server running on port 3001'))
@@ -29,42 +29,45 @@ const io = socket(server, {
 
 io.on('connection', (socket) => {
   console.log('connected')
+
   socket.on('SEND_MESSAGE', (data) => {
     console.log(data)
-    io.emit('RECEIVE_MESSAGE', data);
+    io.to(data.roomID).emit('RECEIVE_MESSAGE', data);
+  })
+
+  socket.on('JOIN', (data) => {
+    socket.join(data.roomID)
+    io.to(data.roomID).emit('RECEIVE_MESSAGE', data);
   })
 })
-
-let datas = []
-
 
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
+app.get('/room', roomRouter)
+app.post('/room', roomRouter)
 
+// const pool = new pg.Pool({
+//   host: "localhost",
+//   database: "artist_chat",
+//   user: "enoki",
+//   port: 5432
+// })
+// pool.connect()
 
+// app.post('/room', function(req, res, next) {
+//   const sqlCreate = "INSERT INTO rooms (name, email, delete_password) VALUES ($1, $2, $3)"
+//   const values = [req.body.name, req.body.email, req.body.deletePassword]
+//   pool.query(sqlCreate, values)
+//   .then()
+//   .catch(e => console.error(e.stack))
+// })
 
-const pool = new pg.Pool({
-  host: "localhost",
-  database: "artist_chat",
-  user: "enoki",
-  port: 5432
-})
-pool.connect()
-
-app.post('/room', function(req, res, next) {
-  const sqlCreate = "INSERT INTO rooms (name, email, delete_password) VALUES ($1, $2, $3)"
-  const values = [req.body.name, req.body.email, req.body.deletePassword]
-  pool.query(sqlCreate, values)
-  .then()
-  .catch(e => console.error(e.stack))
-})
-
-app.get('/room', (req, res, next) => {
-  pool.query("SELECT * FROM rooms", (err, result) => {
-    datas = result.rows
-  })
-  res.json(datas)
-})
+// app.get('/room', (req, res, next) => {
+//   pool.query("SELECT * FROM rooms", (err, result) => {
+//     datas = result.rows
+//     res.json(datas)
+//   })
+// })
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
