+++ db/create_db.sql
@@ -1,0 +1,1 @@
+CREATE DATABASE artist_chat;
