+++ db/create_rooms.sql
@@ -1,0 +1,6 @@
+CREATE TABLE IF NOT EXISTS rooms (
+  id serial,
+  name varchar(255),
+  email varchar(255),
+  delete_password varchar(255)
+);
